requirements.txt
pytesseract==0.3.10
Pillow==10.0.0 